--- packages/util-steps/README.md
@@ -1,10 +1,0 @@
-# DBOS Utility Steps
-
-This is a [DBOS](https://docs.dbos.dev/) [step](https://docs.dbos.dev/typescript/tutorials/step-tutorial) library, providing wrappers for common nondeterministic functions.
-
-These include:
-
-- bcrypt: `bcryptGenSalt`, `bcryptHash`, `bcryptCompare`
-- Date and time: `getCurrentDate`, `getCurrentTime`
-- Random: `random`, `randomInt`
-- UUID generation: `randomUUID`
