--- packages/util-steps/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@dbos-inc/util-steps",
-  "version": "0.0.0-placeholder",
-  "description": "Library of DBOS steps for common non-deterministic functions",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dbos-inc/dbos-transact-ts",
-    "directory": "packages/util-steps"
-  },
-  "homepage": "https://docs.dbos.dev/",
-  "main": "index.js",
-  "scripts": {
-    "build": "tsc --project tsconfig.json",
-    "test": "jest --detectOpenHandles"
-  },
-  "dependencies": {
-    "bcryptjs": "^2.4.3"
-  },
-  "peerDependencies": {
-    "@dbos-inc/dbos-sdk": "*"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.14",
-    "@types/bcryptjs": "^2.4.6",
-    "@types/pg": "^8.15.2",
-    "jest": "^29.7.0",
-    "pg": "^8.11.3",
-    "typescript": "^5.4.5"
-  }
-}
