--- packages/util-steps/tsconfig.json
@@ -1,9 +1,0 @@
-/* Visit https://aka.ms/tsconfig to read more about this file */
-{
-  "extends": "../../tsconfig.shared.json",
-  "compilerOptions": {
-    "outDir": "./dist"
-  },
-  "include": ["index.ts"],
-  "exclude": ["dist", "tests"]
-}
